% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta_ci.R
\name{betaCISample}
\alias{betaCISample}
\title{Sample from the beta distributions}
\usage{
betaCISample(alpha, beta, n = 5000)
}
\arguments{
\item{alpha, beta}{Numeric values with the two shape parameters of the
distribution}

\item{n}{Number of points to be sampled}
}
\description{
Sample from the beta distributions
}
