% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config}
\alias{config}
\title{Sample psiplot configuration settings for datasets \code{psi} and \code{crpkm}}
\format{A 4 x 4 data frame}
\usage{
config
}
\description{
Example of how a psiplot configuration file should be formatted. This can be
passed to \code{\link{plot_event}} or \code{\link{plot_expr}} using the
\code{config} argument.
}
\keyword{datasets}
