% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crpkm}
\alias{crpkm}
\title{Sample genes with cRPKM data}
\format{A 6 x 10 data frame}
\usage{
crpkm
}
\description{
Contains sample cRPKM data as produced by \code{vast-tools combine}.
}
\keyword{datasets}
