% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crpkm_counts}
\alias{crpkm_counts}
\title{Sample genes with cRPKM and counts data}
\format{A 6 x 18 data frame}
\usage{
crpkm_counts
}
\description{
Contains sample cRPKM data and raw read counts
}
\keyword{datasets}
