% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw_group_means.R
\name{draw_group_means}
\alias{draw_group_means}
\title{Draw group means as horizontal lines}
\usage{
draw_group_means(gp, reordered, offset = 2)
}
\arguments{
\item{gp}{ggplot2 object}

\item{reordered}{a list containing input data and plot configuration. Produced
by \code{\link{preprocess_sample_colors}}}

\item{offset}{an integer indicating the offset to print group label from line}
}
\value{
ggplot2 object
}
\description{
When this function is called, the group means are calculated and plotted
as horizontal lines. For internal use only.
}
\seealso{
\code{\link{preprocess_sample_colors}}, \code{\link{plot_event}},
\code{\link{plot_expr}}
}

