% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format_table}
\alias{format_table}
\title{Format input event data}
\usage{
format_table(m)
}
\arguments{
\item{m}{A data frame of event data as outputted by
\code{vast-tools combine}. e.g. each row is an event containing exon
metadata, PSI and quality scores values.}
}
\value{
Data frame with PSI and quality scores. Rownames set as a
concatenation of exon metadata delimited by |.
}
\description{
Convert low/bad quality PSI values and convert event metadata as rownames.
Prepares input event data for plotting. Calls \code{\link{convert_psi}}.
}
\examples{
# For example input, see:
mm.psi
format_table(mm.psi)
}
\seealso{
\code{\link{convert_psi}}
}

