% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta_ci.R
\name{get_beta_ci_subg}
\alias{get_beta_ci_subg}
\title{Calculate conficence intervals for error bars of subgrouped samples.}
\usage{
get_beta_ci_subg(v, q)
}
\arguments{
\item{v}{Numeric vector with PSI values of samples in a subgroup.}

\item{q}{Character vector with the quality scores string of samples in a subgroup.}
}
\description{
Helper function to filter and return confidence intervals based
on a joint beta distribution obtained from Q scores. For internal use.
}
\details{
Individual beta distributions are generated from each sample in the
subgroup. Those distributions are sampled and the results are used to fit a
new joint beta distribution. The joint beta is then sampled again to obtain
confidence intervals for the subgroup.
}
