% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample_colors.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Simulate default ggplot2 color palette}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number of colors}
}
\value{
vector of colors of length n
}
\description{
Simulate default ggplot2 color palette
}
\details{
\url{https://stackoverflow.com/a/8197703}
}
