% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event.R
\name{make_title.2}
\alias{make_title.2}
\title{Make plot title (version 2)}
\usage{
make_title.2(gene, event)
}
\arguments{
\item{gene}{A vector or character string}

\item{event}{A vector or character string with same size as \code{gene}}
}
\value{
A character with a human-friendly title with format:
\code{GENE (EVENT ID)}
}
\description{
Create a plot title using the event ID. Returns symbol and event ID.
}
\examples{
print(make_title.2(psi$GENE, psi$EVENT))
}
