% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mm.psi.config}
\alias{mm.psi.config}
\title{Sample psiplot configuration settings for dataset \code{mm.psi}}
\format{A 4 x 4 data frame}
\usage{
mm.psi.config
}
\description{
Example of how a psiplot configuration file should be formatted. This can be
passed to \code{psiplot::plot_event} using the \code{config} argument, which
accepts either the file path of the config file or an \code{n x 4} data frame where
\code{n} is the number of samples.
}
\keyword{datasets}

