% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_event}
\alias{plot_event}
\title{Plot PSI values of a given alternative splicing event}
\usage{
plot_event(x, config = NULL, errorbar = TRUE,
  groupmean = ifelse(is.null(config), FALSE, TRUE), col = NULL,
  title = NULL, xlab = "", ylab = "PSI", ylim = c(1, 100), xlim = c(1,
  ncol(x)/2), cex.main = 0.9, cex.axis = 0.8, cex.xaxis = 0.6, pch = 20,
  cex.pch = 1, lines = FALSE, gridlines = TRUE)
}
\arguments{
\item{x}{A 1-row data frame containing PSI values to be plotted}

\item{config}{Optional configuration settings for \code{plot_event}. Can be
a path to the \code{.config} file, or 4-column data frame of the \code{.config} file. Use the latter option if you are calling \code{plot_event} multiple times.}

\item{errorbar}{Logical indicating whether error bars should be drawn}

\item{groupmean}{Logical indicating whether grouped means should be drawn.
Requires \code{config}.}

\item{col}{Vector of colors for the points on the plot}

\item{title}{Title of the plot}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{ylim}{Range of y-axis}

\item{xlim}{Range of x-axis}

\item{cex.main}{Plot title size}

\item{cex.axis}{Axis label size}

\item{cex.xaxis}{Size of sample name labels on x-axis}

\item{pch}{Point symbol}

\item{cex.pch}{Size of datapoints}

\item{lines}{Draw a connecting line between points for an event}

\item{gridlines}{Logical indicating whether grid lines should be drawn}
}
\description{
Generate a plot of PSI values for a given exon. The PSI values and
corresponding quality scores are typically obtained from the
\href{https://www.github.com/vastgroup/vast-tools}{vast-tools}
pipeline.
}
\details{
Plots can be customized via the \emph{config} option. Either a data frame or
the filepath to the config file can be used. Alternatively, plots can be
customized using a limited set of graphical parameters as described above.

The order of samples (e.g. columns in \emph{x}) as it appears on the
resulting plot can be customized using a config file. If a config file is not
used and re-ordering is desired, then it must be done manually before calling
\emph{plot_event} by altering the columns of \emph{x}.

If groups are defined in \emph{config} and \emph{groupmean=TRUE}, the mean
PSI of the samples within each group are drawn as horizontal lines. The
colors of the lines are determined by the color set to the first sample of
each group by \emph{RColorCode} in \emph{config}. A corresponding legend key
will also be drawn.

PSI values that have \emph{NA} value are omitted and not plotted.

Error bars based on the confidence interval of the PSI estimation can be
shown by setting \emph{errorbar=TRUE}.
}
\examples{
plot_event(mm.psi[1,])

# Plot with custom configuration
mm.psi.config
plot_event(mm.psi[1,], config = mm.psi.config, groupmean=TRUE)
plot_event(mm.psi[1,], config = "/path/to/mm.psi.config")

# Plot using custom configuration, changing point sympol, and y-axis
# scale
plot_event(mm.psi[1,], config = mm.psi.config, pch = 19, ylim = c(20, 80))
}

