% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.R
\name{plot_multi}
\alias{plot_multi}
\title{Plot multiple events or genes as a heatmap (experimental)}
\usage{
plot_multi(df, config = NULL, expr = FALSE, xlab = "", ylab = "",
  title = "", cluster_rows = TRUE, cluster_cols = ifelse(is.null(config),
  TRUE, FALSE), fill = NULL, usepkg = "ggplot2", ...)
}
\arguments{
\item{df}{A data frame of input values (PSI or cRPKM). If the latter, need to
set \code{expr = TRUE}.}

\item{config}{Optional configuration settings. Can be
a path to the \code{.config} file, or 4-column data frame of the \code{.config}
file.}

\item{expr}{Logical - \code{TRUE} if plotting cRPKMs, \code{FALSE} otherwise}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{title}{Title of the plot}

\item{cluster_rows}{Logical to cluster rows using heirarchical clustering}

\item{cluster_cols}{Logical to cluster columns using heirarchical clustering}

\item{fill}{A vector of colours. e.g. from \code{colorRampPalette}.
Default is \code{NULL}, which will choose the palette automatically.}

\item{usepkg}{Default is \code{pheatmap}, which calls \code{\link[pheatmap]{pheatmap}}.
Other methods are also supported: \code{\link[gplots]{heatmap.2}} and
\code{\link[ggplot2]{geom_tile}}.}

\item{...}{Additional parameters passed to \code{\link[pheatmap]{pheatmap}}
or \code{\link[gplots]{heatmap.2}}}
}
\description{
Visualize multiple PSI or cRPKM values in a single plot.
}
\details{
By default, \code{\link[ggplot2]{geom_tile}} is used to generate a heatmap
and a \code{ggplot2} object is returned. Alternatively, other heatmap
packages are supported including:  \code{\link[pheatmap]{pheatmap}}
(use  \code{usepkgs = "pheatmap"}) and
\code{\link[gplots]{heatmap.2}} (use \code{usepkgs = "gplots"}).

Input is similar to \code{\link{plot_event}} or \code{\link{plot_expr}}.

If \code{cluster_rows = TRUE}, then a heirarchical clustering using \code{hclust}
will be performed on a distance matrix computed by \code{dist}. If \code{config}
is not specified, then the samples will also be clustered.

To set the colours, use \code{fill} option to specify vector of colours.
By default (\code{fill=NULL}), PSI uses a yellow/blue gradient, while cRPKMs
uses the "YlOrRd" color brewer palette.
}
\examples{
# Uses ggplot2 by default
plot_multi(psi)
plot_multi(psi, config = config)

# Use expr = TRUE for cRPKMs
plot_multi(crpkm, expr = TRUE)
plot_multi(crpkm, config = config, expr = TRUE)
plot_multi(crpkm, config = config, expr = TRUE, cluster_rows = TRUE)

# To use pheatmap or gplots for plotting, set usepgk option
plot_multi(psi, config = config, usepkg = "pheatmap")
}
