% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psi}
\alias{psi}
\title{Sample events with PSI data}
\format{A 5 x 22 data frame}
\usage{
psi
}
\description{
Contains sample PSI and quality score data as produced by \code{vast-tools
combine}.
}
\keyword{datasets}
