% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_psi}
\alias{convert_psi}
\title{Convert low/bad quality PSI values to NA}
\usage{
convert_psi(t)
}
\arguments{
\item{t}{Original PSI plus quality scores data frame WITHOUT the exon
metadata columns}
}
\value{
Data frame with the same dimensions as \emph{t} and low/bad quality P
SI values converted to \code{NA}
}
\description{
Helper function to filter and return PSI values.
PSIs are converted to NA if first coverage code is 'N'
e.g. PSI=100, Coverage=N,N,N,OK,S ---> PSI=NA
For internal use. Called by \code{\link{format_table}}.
}
\examples{
convert_psi(mm.psi[,7:ncol(mm.psi)])
}

