% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crpkm}
\alias{crpkm}
\title{Sample genes with simulated cRPKM data}
\format{A 5 x 6 data frame}
\usage{
crpkm
}
\description{
Contains simulated cRPKM data as produced by \code{vast-tools
combine}.
}
\keyword{datasets}

