% Generated by roxygen2 (4.0.2): do not edit by hand
\name{draw_group_means}
\alias{draw_group_means}
\title{Draw group means as horizontal lines}
\usage{
draw_group_means(reordered)
}
\arguments{
\item{reordered}{a list containing input data and plot configuration. Produced
by \code{\link{preprocess_sample_colors}}}
}
\description{
When this function is called, the group means are calculated and plotted
as horizontal lines. For internal use only.
}
\seealso{
\code{\link{preprocess_sample_colors}}, \code{\link{plot_event}},
\code{\link{plot_expr}}
}

