% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw_group_means.R
\name{draw_group_means}
\alias{draw_group_means}
\title{Draw group means as horizontal lines}
\usage{
draw_group_means(gp, values, cfg, group.col, offset = 3)
}
\arguments{
\item{gp}{ggplot2 object}

\item{values}{a melted data frame of PSI/cRPKM values. Usually comes from \code{\link{plot_event}} or \code{\link{plot_expr}}.}

\item{cfg}{configuration as data frame}

\item{group.col}{group colours as computed from the \code{group.col} attribute returned by \code{\link{preprocess_sample_colors}}}

\item{offset}{an integer indicating the offset to print group label from line}

\item{reordered}{a list containing input data and plot configuration. Produced
by \code{\link{preprocess_sample_colors}}}
}
\value{
ggplot2 object
}
\description{
When this function is called, the group means are calculated and plotted
as horizontal lines. For internal use only.
}
\seealso{
\code{\link{preprocess_sample_colors}}, \code{\link{plot_event}},
\code{\link{plot_expr}}
}

