% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_group_means.R
\name{draw_group_means}
\alias{draw_group_means}
\title{Draw group means as horizontal lines}
\usage{
draw_group_means(gp, mdata, reordered)
}
\arguments{
\item{gp}{ggplot2 object}

\item{mdata}{a gathered data frame of PSI/cRPKM values. Usually comes from
\code{\link{plot_event}} or \code{\link{plot_expr}}.}

\item{reordered}{a list with the subgroup and group structure of the samples.
Usually comes from \code{\link{preprocess_sample_colors}}.}
}
\value{
ggplot2 object
}
\description{
When this function is called, the group means are calculated and plotted
as horizontal lines. For internal use only.
}
\seealso{
\code{\link{preprocess_sample_colors}}, \code{\link{plot_event}},
\code{\link{plot_expr}}
}
