% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_beta_ci}
\alias{get_beta_ci}
\title{Calculate confidence intervals for error bars}
\usage{
get_beta_ci(q)
}
\arguments{
\item{q}{a data frame of PSI and corresponding quality values}
}
\value{
Confidence intevals of PSI values
}
\description{
Helper function to filter and return confidence intervals based on beta
distribution from Q scores. For internal use. Beta sampling functions
provided by Tim Sterne-Weiler.
}

