% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psi_samples.R
\name{get_psi_samples}
\alias{get_psi_samples}
\title{Get samples from PSI input table}
\usage{
get_psi_samples(df, config = NULL, value = TRUE)
}
\arguments{
\item{df}{A data frame of PSI input values.}

\item{config}{Config data frame}

\item{value}{If \code{TRUE}, then returns the sample names. Otherwise, return
the indices.}
}
\value{
If \code{value = TRUE}, A character vector consisting of sample names.
Otherwise, a numeric vector of indices.
}
\description{
Get samples from PSI input table
}
\details{
When \code{config} is provided, then the samples from the \emph{SampleName}
column are used to find the indices in the PSI table.
}
\examples{
\dontrun{
get_psi_samples(psi)
}
}
