% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psi_samples.R
\name{get_psi_samples}
\alias{get_psi_samples}
\title{Get samples}
\usage{
get_psi_samples(df, configdf = NULL, value = TRUE)
}
\arguments{
\item{df}{A data frame of PSI input values.}

\item{configdf}{A 4-column data frame.}

\item{value}{If \code{TRUE}, then returns the sample names. Otherwise, return
the indices.}
}
\value{
If \code{value = TRUE}, A character vector consisting of sample names.
Otherwise, a numeric vector of indices.
}
\description{
Get samples from PSI input table
}
\examples{
\dontrun{
#' get_psi_samples(psi)
}
}
