% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event.R
\name{make_title}
\alias{make_title}
\title{Make plot title}
\usage{
make_title(x)
}
\arguments{
\item{x}{A character containing the event ID like
"\code{S|TSPAN6|chrX:99885756-99885863|108}"}
}
\value{
A character with a human-friendly title with
event type, gene symbol, event coordinates, and length
}
\description{
Create a plot title using the event ID
}
\examples{
\dontrun{
f <- format_table(psi)
print(make_title(rownames(f)[1]))
}
}
