% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_event.R
\name{plot_event}
\alias{plot_event}
\title{Plot PSI values of a given alternative splicing event}
\usage{
plot_event(x, trim_colnames = NULL, config = NULL, subg = TRUE,
  errorbar = TRUE, groupmean = ifelse(is.null(config), FALSE, TRUE),
  col = NULL, title = NULL, xlab = "", ylab = "PSI", ylim = c(0, 100),
  cex.main = 14, cex.yaxis = 12, cex.xaxis = 12, pch = 20,
  cex.pch = 3, plot = NULL, gridlines = TRUE)
}
\arguments{
\item{x}{A 1-row data frame containing PSI values to be plotted.}

\item{trim_colnames}{String that must be searched for and trimmed at the end
of every sample column in x.}

\item{config}{Optional configuration settings for \code{plot_event}. Can be
a path to the \code{.config} file, or 4/5-column data frame of the \code{.config}
file. Use the latter option if you are calling \code{plot_event} multiple times.}

\item{subg}{Logical indicating whether samples should be subgrouped for plotting.
This is useful when plotting many samples. Subgroups are decided based on the
\code{SubgroupName} column of the \code{.config} file or data frame. The average
of all samples in a subgruoup is plotted as a single data point. Error bars
are estimated for subgroups by fitting all PSIs in the subgroup to a joint
beta distribution.}

\item{errorbar}{Logical indicating whether error bars should be drawn}

\item{groupmean}{Logical indicating whether grouped means should be drawn.
Requires \code{config}.}

\item{col}{Vector of colors with length matching the number of samples. If
specificed, this will override the color settings specified in \code{config}.}

\item{title}{Title of the plot. If \code{NULL} (default), the title generated by \code{\link{make_title}}.}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{ylim}{Range of y-axis}

\item{cex.main}{Plot title size (pts)}

\item{cex.yaxis}{Y-axis font size (pts)}

\item{cex.xaxis}{X-axis font size (i.e. the sample names) (pts)}

\item{pch}{Point symbol}

\item{cex.pch}{Size of datapoints}

\item{plot}{(deprecated) prints the plot}

\item{gridlines}{Logical indicating whether grid lines should be drawn}
}
\value{
ggplot2 object
}
\description{
Generate a plot of PSI values for a given exon. The PSI values and
corresponding quality scores are typically obtained from the
\href{https://www.github.com/vastgroup/vast-tools}{vast-tools}
pipeline.
}
\details{
Plots can be customized via the \code{config} option. Either a data frame or
the filepath to the config file can be used. Alternatively, plots can be
customized using a limited set of graphical parameters as described above.

The order of samples (e.g. columns in \emph{x}) as it appears on the
resulting plot can be customized using a config file. If a config file is not
used and re-ordering is desired, then it must be done manually before calling
\code{\link{plot_event}} by altering the columns of \emph{x}.

If groups are defined in \code{config} and \code{groupmean=TRUE}, the mean
PSI of the samples within each group are drawn as horizontal lines. The
colors of the lines are determined by the color set to the first sample of
each group by \code{RColorCode} in \code{config}. A corresponding legend key
will also be drawn.

PSI values that have \emph{NA} value are omitted and not plotted.

Error bars based on the confidence interval of the PSI estimation can be
shown by setting \code{errorbar=TRUE}.
}
\examples{
\dontrun{
plot_event(psi[1,])

# Plot with custom configuration
config
plot_event(psi[1,], config = config, groupmean=TRUE)
plot_event(psi[1,], config = "/path/to/config")

# Plot using custom configuration, changing point sympol, and y-axis
# scale
plot_event(psi[1,], config = config, pch = 9, ylim = c(20, 80))
}
}
\seealso{
\code{\link{format_table}} for performing some initial conversion steps of \code{x}
and \code{\link{preprocess_sample_colors}} for pre-processing of \code{x}
using\code{config}
}
