% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_expr.R
\name{plot_expr}
\alias{plot_expr}
\title{Plot cRPKM values of a given gene}
\usage{
plot_expr(x, config = NULL, groupmean = ifelse(is.null(config), FALSE,
  TRUE), col = NULL, title = NULL, xlab = "", ylab = "Expression",
  ylim = NULL, cex.main = 14, cex.yaxis = 12, cex.xaxis = 12,
  pch = 20, cex.pch = 3, plot = TRUE, lines = FALSE, gridlines = TRUE)
}
\arguments{
\item{x}{A 1-row data frame containing cRPKM values to be plotted}

\item{config}{Optional configuration settings for \code{plot_expr}. Can be
a path to the \code{.config} file, or 4-column data frame of the \code{.config}
file. Use the latter option if you are calling \code{plot_expr} multiple times.}

\item{groupmean}{Logical indicating whether grouped means should be drawn.
Requires \code{config}.}

\item{col}{Vector of colors with length matching the number of samples. If
specificed, this will override the color settings specified in \code{config}.}

\item{title}{Title of the plot}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{ylim}{Range of y-axis}

\item{cex.main}{Plot title size (pts)}

\item{cex.yaxis}{Y-axis font size (pts)}

\item{cex.xaxis}{X-axis font size (i.e. the sample names) (pts)}

\item{pch}{Point symbol}

\item{cex.pch}{Size of datapoints}

\item{plot}{logical. If \code{TRUE} (default), the plot is printed.}

\item{lines}{(deprecated) Draw a connecting line between points for an event}

\item{gridlines}{Logical indicating whether grid lines should be drawn}
}
\value{
ggplot2 object
}
\description{
Generate a plot of cRPKM values for a given gene. The cRPKM values are
obtained from the
\href{https://www.github.com/vastgroup/vast-tools}{vast-tools} pipeline.
}
\details{
Like \code{\link{plot_event}}, plots can be customized via the \code{config}
option. Either a data frame or
the filepath to the config file can be used. Alternatively, plots can be
customized using a limited set of graphical parameters as described above.

See Details of \code{\link{plot_event}} for more information on costumizing
plots.

cRPKM values that have \emph{NA} value are omitted and not plotted.
}
\examples{
\dontrun{
plot_expr(crpkm[1,])

# Plot with custom configuration
config
plot_expr(crpkm[1,], config = config, groupmean=TRUE)
plot_expr(crpkm[1,], config = "/path/to/config")

# Plot using custom configuration, changing point sympol, and y-axis
# scale
plot_expr(crpkm[1,], config = config, pch = 9, ylim = c(20, 80))
}
}
\seealso{
\code{\link{format_table}} for performing some initial conversion steps of \code{x}

\code{\link{preprocess_sample_colors}} for pre-processing of \code{x} using
\code{config}
}

