% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_multi}
\alias{plot_multi}
\title{Plot multiple events or genes as a heatmap}
\usage{
plot_multi(df, config = NULL, expr = FALSE, xlab = "", ylab = "",
  title = "Heatmap", cluster_rows = FALSE)
}
\arguments{
\item{df}{A data frame of input values (PSI or cRPKM). If the latter, need to
set \code{expr = TRUE}.}

\item{config}{Optional configuration settings. Can be
a path to the \code{.config} file, or 4-column data frame of the \code{.config}
file.}

\item{expr}{Logical - \code{TRUE} if plotting cRPKMs, \code{FALSE} otherwise}

\item{xlab}{The x-axis label}

\item{ylab}{The y-axis label}

\item{title}{Title of the plot}

\item{cluster_rows}{Logical to cluster rows using heirarchical clustering}
}
\description{
Visualize multiple PSI or cRPKM values in a single plot.
}
\details{
Input is similar to \code{\link{plot_event}} or \code{\link{plot_expr}}.

Uses \code{\link[ggplot2]{geom_tile}} to generate heatmap.
}
\examples{
\dontrun{
plot_multi(psi, config = config)
plot_multi(crpkm, config = config, expr = TRUE)
plot_multi(crpkm, config = config, expr = TRUE, cluster_rows)
}
}

