% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess_sample_colors}
\alias{preprocess_sample_colors}
\title{Preprocess PSI data frame using configuration file}
\usage{
preprocess_sample_colors(psi, config)
}
\arguments{
\item{data}{a \emph{n} x \emph{2*m} data frame of PSI and quality score values where \emph{n} is the number of AS events
and \emph{m} is the number of samples.}

\item{config}{filename of the configuration file for \code{data}. Also
accepts \emph{m'} x \emph{4} data frame of the configuration file, \code{m <=
m}}
}
\value{
A list containing:
 \item{data}{data frame of PSI values with columns re-ordered}
 \item{qual}{data frame of quality scores with columns re-ordered}
 \item{col}{vector of colors corresponding to the re-ordered columns}
 \item{group.index}{list of column indices corresponding to each \emph{GroupName}}
 \item{group.col}{vector of colors corresponding to each \emph{GroupName}}
}
\description{
\code{preprocess_sample_colors} re-orders PSI sample columns to a specified order
and generates a corresponding color code sequence. The sample order and
corresponding colors are taken from a pre-defined tab-delimited file
(see \code{details}).
}
\details{
\code{preprocess_sample_colors} depends on a pre-defined "sample inventory"
database file in tab-delimited format. This file is species-specific and
consists of four columns: \emph{Order, SampleName, GroupName, RColorCode}. The
header is required in the file.

For example:

\preformatted{Order    SampleName    GroupName    RColorCode
1        Ooctye        EarlyDev     blue
2        Embr_2C       EarlyDev     red
etc..}

where:
\itemize{
 \item Order: The ordering of the samples from left to right.
 \item SampleName: Name of the sample. MUST match sample name in input table.
 \item Group name: Use for plotting the average PSI of samples belonging
to the same group (need to use option -u/--group-means)
 \item RColorCode: Color name as specified by \code{colors()} or hex color code (\code{#RRGGBB})
}

The \emph{SampleName} must match the column names in \code{psi}. It is possible
for \code{config} to contain more samples than the \code{psi}. In this case,
the extra samples will be ignored. It is also possible that
\code{config} contains only a subset of the samples in \code{psi}. In this
case, only the samples specified in the \code{config} will be plotted and everything
else is ignored.
}
\examples{
reorderedpsi <- preprocess_sample_colors(mm.psi, config = mm.psi.config)
}

