% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample_colors.R
\name{preprocess_sample_colors}
\alias{preprocess_sample_colors}
\title{Preprocess PSI data frame using configuration file}
\usage{
preprocess_sample_colors(data, config, expr = FALSE, col = NULL)
}
\arguments{
\item{data}{A \emph{n} x \emph{2*m} data frame of PSI and quality score values
where \emph{n} is the number of AS events #' and \emph{m} is the number of samples.}

\item{config}{Filename of the configuration file for \code{data}. Also
accepts \emph{m*} x \emph{4} data frame of the configuration file,
\code{m* <= m}}

\item{expr}{Set to \code{TRUE} if formatting a cRPKM table. Otherwise, \code{FALSE}.}

\item{col}{Vector of colors with length matching the number of samples. If
specificed, this will override the color settings specified in \code{config}.}
}
\value{
A list containing:
\describe{
 \item{data}{data frame of PSI values with columns re-ordered}
 \item{qual}{data frame of quality scores with columns re-ordered. \code{NULL}
 if \code{expr = TRUE}}
 \item{col}{vector of colors corresponding to the re-ordered columns}
 \item{group.index}{list of column indices corresponding to each \code{GroupName}}
 \item{group.col}{vector of colors corresponding to each \code{GroupName}}
 \item{config}{config data frame}
 }
}
\description{
\code{preprocess_sample_colors} re-orders PSI sample columns to a specified order
and generates a corresponding color code sequence. The sample order and
corresponding colors are taken from a pre-defined tab-delimited file
(see \code{details}).
}
\details{
\code{preprocess_sample_colors} depends on a pre-defined "sample inventory"
database file in tab-delimited format. This file is species-specific and
consists of four columns: \emph{Order, SampleName, GroupName, RColorCode}. The
header is required in the file.

For example:

\preformatted{Order    SampleName    GroupName    RColorCode
1        Ooctye        EarlyDev     blue
2        Embr_2C       EarlyDev     red
etc..}

where:
\itemize{
 \item{Order: The ordering of the samples from left to right}
 \item{SampleName: Name of the sample. MUST match sample name in input table}
 \item{Group name: Use for plotting the average PSI of samples belonging
to the same group (need to use option -u/--group-means)}
 \item {RColorCode: Color name as specified by \code{\link{colors}} or hex
 color code (\code{#RRGGBB})}
}

The \emph{SampleName} must match the column names in \code{data}. It is possible
for \code{config} to contain more samples than the \code{data}. In this case,
the extra samples will be ignored. It is also possible that
\code{config} contains only a subset of the samples in \code{data}. In this
case, only the samples specified in the \code{config} will be plotted and everything
else is ignored.

The colors can be overridden by specifiying \code{col}. This was mainly added
to support the \code{col} option provided by \code{\link{plot_event}} --
particularly when \code{config} is not provided.

This function is also used for formatting cRPKM input data by setting
\code{expr = TRUE}.
}
\examples{
reorderedpsi <- preprocess_sample_colors(psi, config = config)

reorderedcrpkm <- preprocess_sample_colors(crpkm, config = config, expr = TRUE)
}
\seealso{
\code{\link{plot_event}}
}

