% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psi}
\alias{psi}
\title{Sample genes with simulated PSI data}
\format{A 5 x 14 data frame}
\usage{
psi
}
\description{
Contains simulated PSI and quality score data as produced by \code{vast-tools
combine}.
}
\keyword{datasets}

