% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{psi}
\alias{psi}
\title{Sample genes with simulated PSI data}
\format{A 5 x 14 data frame}
\usage{
psi
}
\description{
Contains simulated PSI and quality score data as produced by \code{vast-tools
combine} of a few randomly selected mouse genes.
}
\keyword{datasets}

